backend/pyproject.toml
[tool.poetry]
name = "quake-archive-backend"
version = "0.1.0"
description = "QuakeArchive Backend"
authors = ["Vlad Filip"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
numpy = "^1.26.2"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
pandas-geojson = "^1.2.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
